helpers/requirements.txt
climage==0.2.0
Pillow>=10.0.0
